% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_score_get.R
\name{ace_score_get}
\alias{ace_score_get}
\title{Pull highest ACE score from ACE questions}
\usage{
df<-ace_score_get(source, timeframe)
}
\description{
Pull highest ACE score from ACE questions
}
\note{
In accordance with agency practice this query pulls the highest value
recorded for each ACE question. In other words, if a youth indicates
emotional abuse in one interview and no emotional abuse in the next, the
"emotional abuse" question will still be counted toward their final score.

This pulls direct responses from the ACE questions asked as part of the PACT.
These have only recently been implemented, so some older youth are lacking
scores. For historical youth, ACE responses extrapolated from other PACT
questions can be obtained using the argument source = "pact"

For Ace + 10, use the source = "plus ten" to get the ten ACE questions as well
as the additional 10 ACE questions that constitute the expanded assessment.
}
\section{Source}{
 Tables in NSGCMS_RESEARCH, the Noble database.
}

