% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axon_bodycam_log_get_from_api.R
\name{axon_bodycam_log_get_from_api}
\alias{axon_bodycam_log_get_from_api}
\title{Pull formatted youth assignments.}
\usage{
df <- axon_bodycam_log_get_from_api(date, source)
}
\description{
Pull formatted youth assignments.
}
\note{
Use lapply/map for multiple days. Long load time.
}
\section{Source}{
 Queries Axon's evidence.com API to pull the Evidence Created
  report
}

