% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capps_employment_spells.R
\name{capps_employment_spells}
\alias{capps_employment_spells}
\title{Converts formatted CAPPS job history to a summary of all spells of employment
(hire and termination) at TJJD.}
\usage{
df <- capps_employment_spells()
}
\description{
Converts formatted CAPPS job history to a summary of all spells of employment
(hire and termination) at TJJD.
}
\note{
UID is Empl_ID combined with Hire Month.

For name, title, location, position ID, and supervisor position ID, the query
defaults to being effective as of the end of the month. In other words, if a
JCO transfers from RJ to Mart in the month of June, this table will list their
location as being Mart.
}
