% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capps_job_history_get.R
\name{capps_job_history_get}
\alias{capps_job_history_get}
\title{Pull and formats CAPPS job history report prepared by HR}
\usage{
df <- capps_job_history_get()
}
\description{
Pull and formats CAPPS job history report prepared by HR
}
\note{
UID is Empl_ID combined with Start_Date.

Query collapses multiple employee actions on the same day to a single row. In
other words, if an employee had a status change and a pay raise on the same
day, the "Actions" column will include both of these.
}
