% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capps_job_history_monthly.R
\name{capps_job_history_monthly}
\alias{capps_job_history_monthly}
\title{Converts formatted CAPPS job history to a monthly display.}
\usage{
df <- capps_job_history_monthly()
}
\description{
Converts formatted CAPPS job history to a monthly display.
}
\note{
UID is Empl_ID combined with Month.

Query collapses all employee actions into the month in which they occurred. In
other words, if an employee transferred location and received a pay raise on
in the same month, the "Actions" column will include both of these.

For name, title, location, position ID, and supervisor position ID, the query
defaults to being effective as of the end of the month. In other words, if a
JCO transfers from RJ to Mart in the month of June, this table will list their
location as being Mart.
}
