% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capps_leave_by_month.R
\name{capps_leave_by_month}
\alias{capps_leave_by_month}
\title{Converts formatted CAPPS job history to a monthly display.}
\usage{
df <- capps_leave_by_month()
}
\description{
Converts formatted CAPPS job history to a monthly display.
}
\note{
UID is Empl_ID combined with Month.
}
