% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cseit_get.R
\name{cseit_get}
\alias{cseit_get}
\title{Pull formatted records of highest youth CSE-IT (trafficking risk) level}
\usage{
df <- cseit_Get()
}
\description{
Pull formatted records of highest youth CSE-IT (trafficking risk) level
}
\note{
A CSE-IT risk score is shaped largely by fixed factors. This query pulls the
highest CSE-IT risk score in a youth's record.

In other words, if a youth registers as a "clear concern" the first time they
are scored and "no concern" the second time, this query will return the value
"clear concern".

Data filtered on "completed" status.
}
\section{Source}{
 Tables in NSGCMS_RESEARCH, the Noble database.
}

