% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses_060_get.R
\name{diagnoses_060_get}
\alias{diagnoses_060_get}
\title{Pull formatted records of DSM diagnoses on CCF060D}
\usage{
df<-diagnoses_060_Get()
}
\description{
Pull formatted records of DSM diagnoses on CCF060D
}
\note{
Diagnostic list does not include UTMB medical diagnoses, which account for
majority of youth mental health diagnoses
}
\section{Source}{
 Reporting data warehouse copy of Psychological Services Manager tables
}

