% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnoses_utmb_get.R
\name{diagnoses_utmb_get}
\alias{diagnoses_utmb_get}
\title{Pull formatted records of DSM diagnoses in UTMB data}
\usage{
df <- diagnoses_utmb_get()
}
\description{
Pull formatted records of DSM diagnoses in UTMB data
}
\note{
All youth diagnoses made on or after 1/1/2014.

Diagnostic list does not include medical diagnoses on form 060, which account
for a smaller number of youth mental health diagnoses.
}
\section{Source}{
 UTMB Database in IRD035, mostly
  UTMB_DATA_ENCOUNTER_PROC_DIAG
}

