% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee_summary_get.R
\name{employee_summary_get}
\alias{employee_summary_get}
\title{Pull formatted records of employee demographics}
\usage{
df <- employee_summary_get()
}
\description{
Pull formatted records of employee demographics
}
\note{
CCS Location code "MCA" appears in original data when it should be "MCL".
Likewise "HNU" appears when it should be "HKH". This query replaces those
values.

Data filtered on employees with a TINSID (meaning active since 2018 or so).
}
\section{Source}{
 TYC_PG2_JOB_CURR, a table created by IT that summarizes
  basic information of current employees, queried from database
  "TYCEmployees". This is in turn reduced in size into a view
  "vCAPPS_PG2_JOB_CURR" that restricts the query to employees who have been
  entered into CAPPS (and this is the table that the formatted query pulls
  from).

  UID is based on employee ID, meaning that changes over time are not
  accounted for. Each row contains only the most recent data for each
  employee, meaning that if an employee used to be a JCO and is now a case
  manager, their title will only show as "case manager".  Use
  CAPPS_Job_History_Get() for

  In addition, the formatted query pulls rehire dates for currently-employed
  rehires from CAPPS_CEI_Empl also in "TYCEmployees".

  The raw 'sql' version of this query pulls the full TYC_PG2_JOB_CURR and no
  rehire information from CAPPS_CEI_Empl.
}

