% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper functions.R
\name{interval_convert.date}
\alias{interval_convert.date}
\title{Interval Generator}
\usage{
interval_convert.date(
  df,
  start_col,
  end_col,
  int_unit,
  int_length = 1,
  remove_zero_duration = FALSE
)
}
\arguments{
\item{DF}{the dataframe}

\item{Start_Col}{the name of the variable (must be Date or Posix) defining period start}

\item{End_Col}{the name of the variable (must be Date or Posix) defining period end}

\item{Int_Unit}{the time interval used (e.g. "year", "month", "day", hour" "min", "sec". Must be in quotes.)}
}
\description{
Converts tables in "UID - Start Date/time - End Date/time" format to "UID - Interval_Start - Interval Time" format
}
\examples{
Test_DFDate <- data.frame(UID = c(1,2,3),
                         Start_Date = as_date(c("2010-01-01", "2010-04-01", "2011-01-01")),
                         End_Date = as_date(c("2010-06-30", "2011-12-31", "2011-01-02")))

Interval_Convert.Date(Test_DFDate, Start_Date, End_Date, "day", 1)
Interval_Convert.Date(Test_DFDate, Start_Date, End_Date, "month", 1)
Interval_Convert.Date(Test_DFDate, Start_Date, End_Date, "month", 3)
Interval_Convert.Date(Test_DFDate, Start_Date, End_Date, "year", 1)

Test_DFDate_Time <- data.frame(UID = c(1,2,3),
                             Start_Date_Time = as.POSIXct(c("2010-01-01 01:00:00", "2010-01-01 01:00:00", "2010-01-01 01:00:00")),
                             End_Date_Time = as.POSIXct(c("2010-01-01 01:00:10", "2010-01-01 01:10:00", "2010-01-01 02:02:00")))

Interval_Convert.Posix(Test_DFDate_Time, Start_Date_Time, End_Date_Time, "sec", 1)
Interval_Convert.Posix(Test_DFDate_Time, Start_Date_Time, End_Date_Time, "min", 1)
Interval_Convert.Posix(Test_DFDate_Time, Start_Date_Time, End_Date_Time, "hour", 1)
}
