% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pact_risk_factor_get.R
\name{pact_risk_factor_get}
\alias{pact_risk_factor_get}
\title{Pull formatted records of PACT and RPACT risk history for youth}
\usage{
df <- pact_risk_factor_get()
}
\description{
Pull formatted records of PACT and RPACT risk history for youth
}
\note{
This only pulls risk factors, ranks, and their percent of maximum.
}
\section{Source}{
 Noble tables from "NSGCMS_Research".
}

