% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmacy_orders_get.R
\name{pharmacy_orders_get}
\alias{pharmacy_orders_get}
\title{Pull formatted records of pharmacy orders in UTMB data}
\usage{
df <- pharmacy_orders_get()
}
\description{
Pull formatted records of pharmacy orders in UTMB data
}
\note{
Highly truncated (original is 200+ columns) query of drugs prescribed to
youth.

Includes all youth prescriptions in which the first dosage date is on or
after 1/1/2014.

The "START_REASON" column is blank for most entries because it only applies
to non-formulary doses.
}
\section{Source}{
 UTMB Database in IRD035, mostly
  UTMB_DATA_PHARMACY_ORDER_DETAIL
}

